+++ web-ui/src/components/replay-components/ReplayQuarteredEventFrame.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useMemo, useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
-import { EventRecordsApi, EventDtoOutput, EventType } from '../../../generated';
+import { EventRecordsApi, EventDtoOutput } from '../../../generated';
 import { QuarteredEventRecordDtoOutput as EventRecordDto } from '../../../generated/models/QuarteredEventRecordDtoOutput';
 import { getCurrentQuarter, getQuarterEnds } from '../event-components/event-utils';
 import defaultConfig from '../../default-config';
 import { QuarteredEventUI } from '../event-components/quartered-event/QuarteredEventUI';
-
+import { useReplayTickDelay } from '../../hooks/useReplayTickDelay';
 interface ReplayQuarteredEventFrameProps {
   event: EventDtoOutput;
   tick: number;
@@ -15,35 +15,42 @@ export const ReplayQuarteredEventFrame: React.FC<ReplayQuarteredEventFrameProps>
   const [eventRecordsCache, setEventRecordsCache] = useState<Map<number, EventRecordDto[]>>(new Map());
   const [displayedRecords, setDisplayedRecords] = useState<EventRecordDto[]>([]);
   const [currentBlobIndex, setCurrentBlobIndex] = useState(-1);
-  const isOneShot = useMemo(() => event.type === EventType.QuarteredOneShotScoring, [event.type]);
+  const [loadingNextTick, setLoadingNextTick] = useState(false);
   const quarterEnds = useMemo(
-    () => getQuarterEnds(event.competitors.length, isOneShot),
-    [event.competitors.length, isOneShot],
+    () => getQuarterEnds(event.competitors.length, event.type),
+    [event.competitors.length, event.type],
   );
-  const quarter = useMemo(() => getCurrentQuarter(quarterEnds, tick), [tick, quarterEnds]);
+  const [quarter, setQuarter] = useState(getCurrentQuarter(quarterEnds, tick));
 
   const eventRecordsApi = new EventRecordsApi(defaultConfig);
 
-  const { mutate: getEventRecords, isPending } = useMutation<EventRecordDto[], Error, number>({
+  const { mutate: getEventRecords } = useMutation<EventRecordDto[], Error, number>({
     mutationFn: (playbackTick: number) =>
       eventRecordsApi.getQuarteredEventRecordsQuarteredGet({
         eventId: event.id,
         playbackTick,
       }),
     onSuccess: (data, playbackTick) => {
+      setLoadingNextTick(false);
+      setQuarter(getCurrentQuarter(quarterEnds, playbackTick));
       setEventRecordsCache((prev) => new Map(prev.set(playbackTick, data)));
       setDisplayedRecords(data);
     },
   });
 
-  useEffect(() => {
-    if (!eventRecordsCache.has(tick)) {
-      getEventRecords(tick);
-    } else {
-      // If we have cached data for this tick, update displayed records
-      setDisplayedRecords(eventRecordsCache.get(tick)!);
-    }
-  }, [tick, eventRecordsCache, getEventRecords]);
+  useReplayTickDelay(
+    tick,
+    () => {
+      if (!eventRecordsCache.has(tick)) {
+        getEventRecords(tick);
+      } else {
+        setDisplayedRecords(eventRecordsCache.get(tick)!);
+        setLoadingNextTick(false);
+        setQuarter(getCurrentQuarter(quarterEnds, tick));
+      }
+    },
+    setLoadingNextTick,
+  );
 
   useEffect(() => {
     if (displayedRecords.length > 0) {
@@ -58,7 +65,7 @@ export const ReplayQuarteredEventFrame: React.FC<ReplayQuarteredEventFrameProps>
       isEventFinished={false}
       eventType={event.type}
       currentBlobIndex={currentBlobIndex}
-      isPerforming={isPending}
+      isPerforming={loadingNextTick}
     />
   );
 };